pyproject.toml
[project]
name = "youtube-playlist-editor"
version = "0.1.0"
description = "A tool to add YouTube videos from a file to a playlist."
authors = [
    {name = "William Ko",email = "4422233+WhiSKill@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client (>=2.166.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "click (>=8.1.8,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"  # Use a recent stable version
pytest-mock = "^3.12.0"
pytest-cov = "^5.0.0"
